Remove some active but useless lines from MMRP
The code removed seems to be leftover copy-paste mess from MVRP;
it is incorrect in MMRP, but the correct operation is done at the
beginning of the function and the leftover lines don't have any
ill effect due to the NULL check in the called code.

There should be no runtime behavioral change due to this.